♻️ Stream the branding settings below a Suspense boundary
The page shell (header) renders synchronously; the branding read moved
into a private async component under Suspense, per the streaming rules.
The section lives in its own file because the i18n scanner only
recognizes the Trans component by its unaliased name, and the shell and
section need different Trans imports.
